Implementação de Exceptions personalizadas, Criação de ExceptionHandler, modificações nos métodos CRUD com interface JPARepository e inclusão de tratamento de exceções neles

+++ src/main/java/one/digitalinnovation/gof/service/impl/ClienteServiceImpl.java
@@ -2,6 +2,8 @@ package one.digitalinnovation.gof.service.impl;
 
 import java.util.Optional;
 
+import javax.persistence.EntityNotFoundException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +13,7 @@ import one.digitalinnovation.gof.model.Endereco;
 import one.digitalinnovation.gof.model.EnderecoRepository;
 import one.digitalinnovation.gof.service.ClienteService;
 import one.digitalinnovation.gof.service.ViaCepService;
+import one.digitalinnovation.gof.service.exceptions.ClienteNotfoundException;
 
 /**
  * Implementação da <b>Strategy</b> {@link ClienteService}, a qual pode ser
@@ -42,31 +45,48 @@ public class ClienteServiceImpl implements ClienteService {
 	@Override
 	public Cliente buscarPorId(Long id) {
 		// Buscar Cliente por ID.
-		Optional<Cliente> cliente = clienteRepository.findById(id);
-		return cliente.get();
+		Optional<Cliente> objCliente = clienteRepository.findById(id);
+		Cliente cliente = objCliente.orElseThrow(() -> new ClienteNotfoundException("Cliente não encontrado"));
+		return cliente;
 	}
 
 	@Override
-	public void inserir(Cliente cliente) {
-		salvarClienteComCep(cliente);
+	public Cliente inserir(Cliente objCliente) {
+		Cliente cliente = checarCepExistente(objCliente);
+		return clienteRepository.save(cliente);
 	}
 
 	@Override
-	public void atualizar(Long id, Cliente cliente) {
-		// Buscar Cliente por ID, caso exista:
-		Optional<Cliente> clienteBd = clienteRepository.findById(id);
-		if (clienteBd.isPresent()) {
-			salvarClienteComCep(cliente);
+	public Cliente atualizar(Long id, Cliente clienteAtualizado) {		
+		try {
+			// Buscar Cliente por ID, caso exista:
+			Cliente cliente = clienteRepository.getById(id);
+			// Checa se o nome do cliente foi alterado
+			if(cliente.getNome() != clienteAtualizado.getNome()) 
+				cliente.setNome(clienteAtualizado.getNome());
+			// Checa se o endereco do cliente foi alterado
+			if(cliente.getEndereco() != clienteAtualizado.getEndereco()) 
+				cliente.setEndereco(checarCepExistente(clienteAtualizado).getEndereco());
+						
+			cliente = clienteRepository.save(cliente);
+			return cliente;
+			
+		} catch (EntityNotFoundException  e) {
+			throw new ClienteNotfoundException("Cliente não encontrado");
 		}
+		
 	}
 
 	@Override
-	public void deletar(Long id) {
-		// Deletar Cliente por ID.
-		clienteRepository.deleteById(id);
+	public void deletar(Long id) {		
+		// Checa existência do cliente
+		if(!clienteRepository.existsById(id))
+			throw new ClienteNotfoundException("Cliente não encontrado");
+		else
+			clienteRepository.deleteById(id); // Deletar Cliente por ID.
 	}
 
-	private void salvarClienteComCep(Cliente cliente) {
+	private Cliente checarCepExistente(Cliente cliente) {
 		// Verificar se o Endereco do Cliente já existe (pelo CEP).
 		String cep = cliente.getEndereco().getCep();
 		Endereco endereco = enderecoRepository.findById(cep).orElseGet(() -> {
@@ -76,8 +96,8 @@ public class ClienteServiceImpl implements ClienteService {
 			return novoEndereco;
 		});
 		cliente.setEndereco(endereco);
-		// Inserir Cliente, vinculando o Endereco (novo ou existente).
-		clienteRepository.save(cliente);
+		// Retorna Cliente, vinculando o Endereco (novo ou existente).
+		return cliente;
 	}
 
 }
